external/webkit/LayoutTests/platform/android/webgl-conformance/dumpAsText/conformance/glsl/misc/shader-with-define-line-continuation.frag-expected.txt

CONSOLE MESSAGE: line 24: *** Error loading shader '[object WebGLShader]':INVALID_VALUE
fragment shader that uses line continuation macro should fail

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

test: fragment shader that uses line continuation macro should fail
default vertex shader
test fragment shaderPASS fragment shader that uses line continuation macro should fail
PASS successfullyParsed is true

TEST COMPLETE
